Move wrap tests to submodule
